// src/models/interface.ts
export type Id = mongoose.Types.ObjectId;
// ...
export interface InterCampBack {
  nameId: Id;
  round: number;
  dateStart: Date;
  dateEnd: Date;
  boardIds: Id[];
  peeIds: Id[];
  nongIds: Id[];
  partIds: Id[];
  petoIds: Id[];
  authorizeIds: Id[];
  nongDataLock: boolean;
  nongModelIds: Id[];
  peeModelIds: Id[];
  petoModelIds: Id[];
  nongPendingIds: Map<Id, string>;
  nongPassIds: Map<Id, string>;
  open: boolean;
  peePassIds: Map<Id, Id>;
  songIds: Id[];
  nongSureIds: Id[];
  baanIds: Id[];
  link: string | null;
  allDone: boolean;
  lockChangePickup: boolean;
  pictureUrls: string[];
  campStyleId: Id;
  actionPlanIds: Id[];
  workItemIds: Id[];
  nongPaidIds: Id[];
  nongInterviewIds: Map<Id, string>;
  registerModel: "noPaid" | "noInterview" | "all";
  memberStructure:
    | "nong->highSchool,pee->1year,peto->2upYear"
    | "nong->highSchool,pee->2upYear"
    | "nong->1year,pee->2upYear"
    | "nong->highSchool,pee->allYear"
    | "allYearMix";
  logoUrl: string | null;
  mapCampMemberCardIdByUserId: Map<Id, Id>;
  registerSheetLink: string | null;
  peeLock: boolean;
  outRoundIds: Id[];
  _id: Id;
  campName: string;
  nongSleepModel:
    | "นอนทุกคน"
    | "เลือกได้ว่าจะค้างคืนหรือไม่"
    | "ไม่มีการค้างคืน";
  peeSleepModel: "นอนทุกคน" | "เลือกได้ว่าจะค้างคืนหรือไม่" | "ไม่มีการค้างคืน";
  groupRefMap: Map<Group, Id>;
  baanBoardId: Id | null;
  partNameIds: Id[];
  partBoardId: Id;
  partPeeBaanId: Id;
  groupName: string;
  peeDataLock: boolean;
  petoDataLock: boolean;
  haveCloth: boolean;
  actionPlanOffset: number;
  nongMapIdLtoG: Map<number, Id>;
  peeMapIdLtoG: Map<number, Id>;
  nongMapIdGtoL: Map<Id, number>;
  peeMapIdGtoL: Map<Id, number>;
  currentNong: number;
  currentPee: number;
  mdTime: Date;
  allPetoChatIds: Id[];
  nongCampMemberCardHaveHeathIssueIds: Id[];
  peeCampMemberCardHaveHeathIssueIds: Id[];
  petoCampMemberCardHaveHeathIssueIds: Id[];
  choiceQuestionIds: Id[];
  textQuestionIds: Id[];
  nongAnswerPackIds: Id[];
  peeAnswerPackIds: Id[];
  mapAnswerPackIdByUserId: Map<Id, Id>;
  peeAnswerIds: Id[];
  showCorrectAnswerAndScore: boolean;
  canAnswerTheQuestion: boolean;
  mealIds: Id[];
  foodIds: Id[];
  canNongSeeAllAnswer: boolean;
  canNongSeeAllActionPlan: boolean;
  canNongSeeAllTrackingSheet: boolean;
  canNongAccessDataWithRoleNong: boolean;
  lockChangeQuestion: boolean;
  jobIds: Id[];
  canReadTimeOnMirror: boolean;
  nongCall: string;
  boyZoneLadyZoneState: BoyZoneLadyZoneState;
  canNongSeeBaanOrder: boolean;
  scoreIds: Id[];
  //public
}
// ...
export interface InterSong {
  name: string;
  campIds: Id[];
  baanIds: Id[];
  author: string;
  time: number;
  link: string;
  userLikeIds: Id[];
  _id: Id;
  //public
}
// ...
export type Group =
  | "A"
  | "B"
  | "C"
  | "Dog"
  | "E"
  | "F"
  | "G"
  | "H"
  | "J"
  | "K"
  | "L"
  | "M"
  | "N"
  | "P"
  | "Q"
  | "R"
  | "S"
  | "T";
// ...
export interface GetMenuSongs {
  songs: ShowSong[];
  likeSongIds: Id[];
  authBaans: {
    data: BasicBaan;
    showName: string;
  }[];
  authCamps: BasicCamp[];
  //private
}
export interface ShowSong {
  name: string;
  campNames: string[];
  baanNames: string[];
  author: string;
  time: number;
  link: string;
  like: number;
  _id: Id;
  baanRelates: string[];
  campRelates: string[];
  //private
}
export interface CreateSong {
  name: string;
  author: string;
  time: number;
  link: string;
  //public
}
export interface ShowSongPage {
  song: ShowSong;
  authBaans: {
    data: BasicBaan;
    showName: string;
  }[];
  authCamps: BasicCamp[];
  likeSongIds: Id[];
  //private
}
export interface UpdateSongs {
  _id: Id;
  songIds: Id[];
  //public
}
export interface UpdateSongPage {
  userLikeSongIds: Id[];
  baans: UpdateSongs[];
  camps: UpdateSongs[];
  //private
}
export interface SongCount {
  song: InterSong;
  count: number;
}
export interface ShowCampSong {
  nongLike: number;
  peeLike: number;
  petoLike: number;
  name: string;
  campNames: string[];
  baanNames: string[];
  author: string;
  time: number;
  link: string;
  like: number;
  _id: Id;
  //public
}
export interface AuthSongsCamp {
  camp: BasicCamp;
  baans: BasicBaan[];
  authCamp: boolean;
  songs: ShowCampSong[];
  userLikeSongIds: Id[];
}
export interface ShowCampSongReady {
  _id: Id;
  songIds: Id[];
  userLikeSongIds: Id[];
  showCampSongs: ShowCampSong[];
  camp: BasicCamp;
  baan: BasicBaan;
  //private
}
export interface BasicUser {
  _id: Id;
  name: string;
  lastname: string;
  nickname: string;
  gender: "Male" | "Female";
  shirtSize: Size;
  haveBottle: boolean;
  mode: Mode;
  group: Group | null;
  role: Role;
  fridayActEn: boolean;
  fridayAuth: boolean;
  likeToSleepAtCamp: boolean;
  //private
}
export interface BasicBaan {
  name: string;
  fullName: string | null;
  campId: Id;
  link: string | null;
  _id: Id;
  mdTime: Date;
  nongSendMessage: boolean;
  songIds: Id[];
  canReadMirror: boolean;
  canWriteMirror: boolean;
  mirrorIds: Id[];
  nongIds: Id[];
  peeIds: Id[];
  //public
}
export interface BasicCamp {
  nameId: Id;
  round: number;
  dateStart: Date;
  dateEnd: Date;
  nongDataLock: boolean;
  open: boolean;
  songIds: Id[];
  link: string | null;
  allDone: boolean;
  lockChangePickup: boolean;
  pictureUrls: string[];
  registerModel: "noPaid" | "noInterview" | "all";
  memberStructure:
    | "nong->highSchool,pee->1year,peto->2upYear"
    | "nong->highSchool,pee->2upYear"
    | "nong->1year,pee->2upYear"
    | "nong->highSchool,pee->allYear"
    | "allYearMix";
  logoUrl: string | null;
  registerSheetLink: string | null;
  peeLock: boolean;
  _id: Id;
  campName: string;
  nongSleepModel:
    | "นอนทุกคน"
    | "เลือกได้ว่าจะค้างคืนหรือไม่"
    | "ไม่มีการค้างคืน";
  peeSleepModel: "นอนทุกคน" | "เลือกได้ว่าจะค้างคืนหรือไม่" | "ไม่มีการค้างคืน";
  groupName: string;
  peeDataLock: boolean;
  petoDataLock: boolean;
  haveCloth: boolean;
  actionPlanOffset: number;
  mdTime: Date;
  showCorrectAnswerAndScore: boolean;
  canAnswerTheQuestion: boolean;
  mealIds: Id[];
  canNongSeeAllAnswer: boolean;
  canNongSeeAllActionPlan: boolean;
  canNongSeeAllTrackingSheet: boolean;
  canNongAccessDataWithRoleNong: boolean;
  lockChangeQuestion: boolean;
  canReadTimeOnMirror: boolean;
  nongCall: string;
  boyZoneLadyZoneState: BoyZoneLadyZoneState;
  canNongSeeBaanOrder: boolean;
  nongIds: Id[];
  peeIds: Id[];
  petoIds: Id[];
  boardIds: Id[];
  baanIds: Id[];
  partIds: Id[];
  //public
}
// ...
export const boyZoneLadyZoneStates = [
  "ปิดสมบูรณ์",
  "เพศตรงข้ามออกจากโซน",
  "ตรวจตรา",
  "พร้อมอาบน้ำ",
  "เปิดสมบูรณ์",
  "ปิดแต่ยังเก็บของยังไม่หมด",
] as const;
export type BoyZoneLadyZoneState = (typeof boyZoneLadyZoneStates)[number];